Layout of carousel componenets created
- Refactored Carousel component into it's own directory
